Don't reinvent the wheel

+++ src/Day1.hs
@@ -1,12 +1,13 @@
 module Day1 where
 
-import Common (Parser, integer, parseFromFile)
+import Common (Parser, parseFromFile)
 import Data.List (sort)
 import Text.Megaparsec (MonadParsec (eof, lookAhead, try), sepBy1)
 import Text.Megaparsec.Char (digitChar, newline)
+import Text.Megaparsec.Char.Lexer (decimal)
 
 integerGroupParser :: Parser [Integer]
-integerGroupParser = sepBy1 integer (try (newline >> lookAhead digitChar))
+integerGroupParser = sepBy1 decimal (try (newline >> lookAhead digitChar))
 
 inputParser :: Parser [[Integer]]
 inputParser = sepBy1 integerGroupParser (try (newline >> newline)) <* newline <* eof
